+++ packages/plugins/GeoLocation/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## unpublished
+
+- Feature: Adding optional tooltip to user geolocation map icon.
+
 ## 1.0.0
 
 Initial release.