+++ packages/types/custom/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## unpublished
+
+- Feature: Add flag `showTooltip` to `GeoLocationConfiguration`.
+
 ## 1.0.0
 
 Initial release.